# tmf676/src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  description: |-
    ## TMF API Reference : TMF 676 - Payment

    This API provides the standardized client interface to Payment Systems for managing performed payments or refunds. Examples of Payment API originators (clients) include Web servers, mobile app servers, Contact center dashboards or retail store systems.
  title: Payment
  version: 4.0.0
servers:
- url: https://serverRootserverRoot/payment/v4/
...
paths:
...
  /refund:
    get:
      description: This operation list or find Refund entities
      operationId: listRefund
      parameters:
      - description: Comma-separated properties to be provided in response
        in: query
        name: fields
        schema:
          type: string
      - description: Requested index for start of resources to be provided in response
        in: query
        name: offset
        schema:
          type: integer
      - description: Requested number of resources to be provided in response
        in: query
        name: limit
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json;charset=utf-8:
              schema:
                items:
                  $ref: '#/components/schemas/Refund'
                type: array
          description: Success
          headers:
            X-Total-Count:
              description: Total number of items matching criteria
              schema:
                type: integer
            X-Result-Count:
              description: Actual number of items returned in the response body
              schema:
                type: integer
        "400":
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad Request
        "401":
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "403":
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden
        "404":
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
          description: Not Found
        "405":
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
          description: Method Not allowed
        "409":
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
          description: Conflict
        "500":
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
          description: Internal Server Error
      summary: List or find Refund objects
      tags:
      - refund
      x-accepts: application/json;charset=utf-8
      x-tags:
      - tag: refund
    post:
      description: This operation creates a Refund entity.
      operationId: createRefund
      requestBody:
        content:
          application/json;charset=utf-8:
            schema:
              $ref: '#/components/schemas/Refund_Create'
        description: The Refund to be created
        required: true
      responses:
        "201":
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Refund'
          description: Created
        "400":
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad Request
        "401":
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "403":
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden
        "405":
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
          description: Method Not allowed
        "409":
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
          description: Conflict
        "500":
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
          description: Internal Server Error
      summary: Creates a Refund
      tags:
      - refund
      x-codegen-request-body-name: refund
      x-content-type: application/json;charset=utf-8
      x-accepts: application/json;charset=utf-8
      x-tags:
      - tag: refund
  /refund/{id}:
    get:
      description: This operation retrieves a Refund entity. Attribute selection is
        enabled for all first level attributes.
      operationId: retrieveRefund
      parameters:
      - description: Identifier of the Refund
        in: path
        name: id
        required: true
        schema:
          type: string
      - description: Comma-separated properties to provide in response
        in: query
        name: fields
        schema:
          type: string
      responses:
        "200":
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Refund'
          description: Success
        "400":
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad Request
        "401":
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "403":
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden
        "404":
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
          description: Not Found
        "405":
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
          description: Method Not allowed
        "409":
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
          description: Conflict
        "500":
          content:
            application/json;charset=utf-8:
              schema:
                $ref: '#/components/schemas/Error'
          description: Internal Server Error
      summary: Retrieves a Refund by ID
      tags:
      - refund
      x-accepts: application/json;charset=utf-8
      x-tags:
      - tag: refund
...
components:
  schemas:
    AccountRef:
      description: Account reference. A account may be a party account or a financial
        account.
      example:
        '@referredType': '@referredType'
        '@baseType': '@baseType'
        '@type': '@type'
        name: name
        description: description
        id: id
        href: href
        '@schemaLocation': https://openapi-generator.tech
      properties:
        id:
          description: Unique identifier of the account
          type: string
        href:
          description: Reference of the account
          type: string
        description:
          description: Detailed description of the account
          type: string
        name:
          description: Name of the account
          type: string
        '@baseType':
          description: "When sub-classing, this defines the super-class"
          type: string
        '@schemaLocation':
          description: A URI to a JSON-Schema file that defines additional attributes
            and relationships
          format: uri
          type: string
        '@type':
          description: "When sub-classing, this defines the sub-class entity name"
          type: string
        '@referredType':
          description: The actual type of the target instance when needed for disambiguation.
          type: string
      required:
      - id
      type: object
    ChannelRef:
      description: "The channel to which the resource reference to. e.g. channel for\
        \ selling product offerings, channel for opening a trouble ticket etc.."
      example:
        '@referredType': '@referredType'
        '@baseType': '@baseType'
        '@type': '@type'
        name: name
        id: id
        href: href
        '@schemaLocation': https://openapi-generator.tech
      properties:
        id:
          description: Unique identifier of a related entity.
          type: string
        href:
          description: Reference of the related entity.
          type: string
        name:
          description: Name of the channel.
          type: string
        '@baseType':
          description: "When sub-classing, this defines the super-class"
          type: string
        '@schemaLocation':
          description: A URI to a JSON-Schema file that defines additional attributes
            and relationships
          format: uri
          type: string
        '@type':
          description: "When sub-classing, this defines the sub-class entity name"
          type: string
        '@referredType':
          description: The actual type of the target instance when needed for disambiguation.
          type: string
      required:
      - id
      type: object
...
    Money:
      description: A base / value business entity used to represent money
      example:
        unit: unit
        value: 0.8008282
      properties:
        unit:
          description: Currency (ISO4217 norm uses 3 letters to define the currency)
          type: string
        value:
          description: A positive floating point number
          format: float
          type: number
      type: object
...
    PaymentMethodRefOrValue:
      description: link to the resource that holds information about the payment mean
        used to complete the operation
      example:
        statusDate: 2000-01-23T04:56:07.000+00:00
        '@referredType': '@referredType'
        validFor:
          startDateTime: 2000-01-23T04:56:07.000+00:00
          endDateTime: 2000-01-23T04:56:07.000+00:00
        '@type': '@type'
        description: description
        isPreferred: true
        relatedParty:
          '@referredType': '@referredType'
          role: role
          '@baseType': '@baseType'
          '@type': '@type'
          name: name
          id: id
          href: href
          '@schemaLocation': https://openapi-generator.tech
        '@baseType': '@baseType'
        name: name
        id: id
        href: href
        '@schemaLocation': https://openapi-generator.tech
        account:
        - '@referredType': '@referredType'
          '@baseType': '@baseType'
          '@type': '@type'
          name: name
          description: description
          id: id
          href: href
          '@schemaLocation': https://openapi-generator.tech
        - '@referredType': '@referredType'
          '@baseType': '@baseType'
          '@type': '@type'
          name: name
          description: description
          id: id
          href: href
          '@schemaLocation': https://openapi-generator.tech
        status: status
      properties:
        id:
          description: Unique Identifier within the server for the payment method.
          type: string
        href:
          description: A resource URI pointing to the resource in the server that
            stores the detailed information. This is typically the resource url to
            retrieve individual details for the specific payment method
          type: string
        description:
          description: Text describing the contents of the payment method
          type: string
        isPreferred:
          description: If the method is the preferred one by the owner. Typically
            used when querying for the payment methods of a specific customer or account
          type: boolean
        name:
          description: Friendly name assigned to the payment method
          type: string
        status:
          type: string
        statusDate:
          format: date-time
          type: string
        account:
          items:
            $ref: '#/components/schemas/AccountRef'
          type: array
        relatedParty:
          $ref: '#/components/schemas/RelatedParty'
        validFor:
          $ref: '#/components/schemas/TimePeriod'
        '@baseType':
          description: "When sub-classing, this defines the super-class"
          type: string
        '@schemaLocation':
          description: A URI to a JSON-Schema file that defines additional attributes
            and relationships
          format: uri
          type: string
        '@type':
          description: "When sub-classing, this defines the sub-class entity name"
          type: string
        '@referredType':
          description: The actual type of the target instance when needed for disambiguation.
          type: string
      type: object
    PaymentRef:
      description: "If an immediate payment has been done at the product order submission,\
        \ the payment information are captured and stored (as a reference) in the\
        \ order."
      example:
        '@referredType': '@referredType'
        '@baseType': '@baseType'
        '@type': '@type'
        name: name
        id: id
        href: href
        '@schemaLocation': https://openapi-generator.tech
      properties:
        id:
          description: Unique identifier of a related entity.
          type: string
        href:
          description: Reference of the related entity.
          type: string
        name:
          description: A name for the payment
          type: string
        '@baseType':
          description: "When sub-classing, this defines the super-class"
          type: string
        '@schemaLocation':
          description: A URI to a JSON-Schema file that defines additional attributes
            and relationships
          format: uri
          type: string
        '@type':
          description: "When sub-classing, this defines the sub-class entity name"
          type: string
        '@referredType':
          description: The actual type of the target instance when needed for disambiguation.
          type: string
      required:
      - id
      type: object
...
    Refund:
      description: The Refund resource represents a performed Refund. It contains
        both information about the refund and the payment method used to perform it.
      example:
        statusDate: 2000-01-23T04:56:07.000+00:00
        amount:
          unit: unit
          value: 0.8008282
        authorizationCode: authorizationCode
        '@type': '@type'
        channel:
          '@referredType': '@referredType'
          '@baseType': '@baseType'
          '@type': '@type'
          name: name
          id: id
          href: href
          '@schemaLocation': https://openapi-generator.tech
        description: description
        requestor:
          '@referredType': '@referredType'
          role: role
          '@baseType': '@baseType'
          '@type': '@type'
          name: name
          id: id
          href: href
          '@schemaLocation': https://openapi-generator.tech
        totalAmount:
          unit: unit
          value: 0.8008282
        '@baseType': '@baseType'
        name: name
        paymentMethod:
          statusDate: 2000-01-23T04:56:07.000+00:00
          '@referredType': '@referredType'
          validFor:
            startDateTime: 2000-01-23T04:56:07.000+00:00
            endDateTime: 2000-01-23T04:56:07.000+00:00
          '@type': '@type'
          description: description
          isPreferred: true
          relatedParty:
            '@referredType': '@referredType'
            role: role
            '@baseType': '@baseType'
            '@type': '@type'
            name: name
            id: id
            href: href
            '@schemaLocation': https://openapi-generator.tech
          '@baseType': '@baseType'
          name: name
          id: id
          href: href
          '@schemaLocation': https://openapi-generator.tech
          account:
          - '@referredType': '@referredType'
            '@baseType': '@baseType'
            '@type': '@type'
            name: name
            description: description
            id: id
            href: href
            '@schemaLocation': https://openapi-generator.tech
          - '@referredType': '@referredType'
            '@baseType': '@baseType'
            '@type': '@type'
            name: name
            description: description
            id: id
            href: href
            '@schemaLocation': https://openapi-generator.tech
          status: status
        correlatorId: correlatorId
        payment:
          '@referredType': '@referredType'
          '@baseType': '@baseType'
          '@type': '@type'
          name: name
          id: id
          href: href
          '@schemaLocation': https://openapi-generator.tech
        id: id
        href: href
        taxAmount:
          unit: unit
          value: 0.8008282
        '@schemaLocation': https://openapi-generator.tech
        refundDate: 2000-01-23T04:56:07.000+00:00
        account:
          '@referredType': '@referredType'
          '@baseType': '@baseType'
          '@type': '@type'
          name: name
          description: description
          id: id
          href: href
          '@schemaLocation': https://openapi-generator.tech
        status: status
      properties:
        id:
          description: Unique identifier of Refund
          type: string
        href:
          description: Hypertext Reference of the refund
          type: string
        authorizationCode:
          description: Authorization code retrieved from an external payment gateway
            that could be used for conciliation
          type: string
        correlatorId:
          description: Unique identifier in the client for the refund in case it is
            needed to correlate
          type: string
        description:
          description: Text describing the contents of the refund
          type: string
        name:
          description: Screen name of the refund
          type: string
        refundDate:
          description: Date when the refund was performed
          format: date-time
          type: string
        status:
          description: Status of the refund
          type: string
        statusDate:
          description: Date when the status was recorded
          format: date-time
          type: string
        account:
          $ref: '#/components/schemas/AccountRef'
        amount:
          $ref: '#/components/schemas/Money'
        channel:
          $ref: '#/components/schemas/ChannelRef'
        payment:
          $ref: '#/components/schemas/PaymentRef'
        paymentMethod:
          $ref: '#/components/schemas/PaymentMethodRefOrValue'
        requestor:
          $ref: '#/components/schemas/RelatedParty'
        taxAmount:
          $ref: '#/components/schemas/Money'
        totalAmount:
          $ref: '#/components/schemas/Money'
        '@baseType':
          description: "When sub-classing, this defines the super-class"
          type: string
        '@schemaLocation':
          description: A URI to a JSON-Schema file that defines additional attributes
            and relationships
          format: uri
          type: string
        '@type':
          description: "When sub-classing, this defines the sub-class entity name"
          type: string
      required:
      - account
      - paymentMethod
      type: object
    Refund_Create:
      description: |-
        The Refund resource represents a performed Refund. It contains both information about the refund and the payment method used to perform it.
        Skipped properties: id,href,status,statusDate,RefundDate
      properties:
        authorizationCode:
          description: Authorization code retrieved from an external payment gateway
            that could be used for conciliation
          type: string
        correlatorId:
          description: Unique identifier in the client for the refund in case it is
            needed to correlate
          type: string
        description:
          description: Text describing the contents of the refund
          type: string
        name:
          description: Screen name of the refund
          type: string
        refundDate:
          description: Date when the refund was performed
          format: date-time
          type: string
        account:
          $ref: '#/components/schemas/AccountRef'
        amount:
          $ref: '#/components/schemas/Money'
        channel:
          $ref: '#/components/schemas/ChannelRef'
        payment:
          $ref: '#/components/schemas/PaymentRef'
        paymentMethod:
          $ref: '#/components/schemas/PaymentMethodRefOrValue'
        requestor:
          $ref: '#/components/schemas/RelatedParty'
        taxAmount:
          $ref: '#/components/schemas/Money'
        totalAmount:
          $ref: '#/components/schemas/Money'
        '@baseType':
          description: "When sub-classing, this defines the super-class"
          type: string
        '@schemaLocation':
          description: A URI to a JSON-Schema file that defines additional attributes
            and relationships
          format: uri
          type: string
        '@type':
          description: "When sub-classing, this defines the sub-class entity name"
          type: string
      required:
      - account
      - paymentMethod
      - totalAmount
      type: object
...
    Error:
      description: "Used when an API throws an Error, typically with a HTTP error\
        \ response-code (3xx, 4xx, 5xx)"
      example:
        reason: reason
        code: code
        '@baseType': '@baseType'
        '@type': '@type'
        message: message
        referenceError: https://openapi-generator.tech
        '@schemaLocation': https://openapi-generator.tech
        status: status
      properties:
        code:
          description: "Application relevant detail, defined in the API or a common\
            \ list."
          type: string
        reason:
          description: Explanation of the reason for the error which can be shown
            to a client user.
          type: string
        message:
          description: More details and corrective actions related to the error which
            can be shown to a client user.
          type: string
        status:
          description: HTTP Error code extension
          type: string
        referenceError:
          description: URI of documentation describing the error.
          format: uri
          type: string
        '@baseType':
          description: "When sub-classing, this defines the super-class."
          type: string
        '@schemaLocation':
          description: A URI to a JSON-Schema file that defines additional attributes
            and relationships
          format: uri
          type: string
        '@type':
          description: "When sub-classing, this defines the sub-class entity name."
          type: string
      required:
      - code
      - reason
      type: object
...